Ported to work with ghc 7.10.
In the next release of GHC, Applicative is becoming a superclass of
Monad. As such, any type which has a Monad instance but not an
Applicative instance will no longer compile. Because of this, an
Applicative instance ahs been added for the ElementFormatter and Blpapi
types.

See
https://www.haskell.org/haskellwiki/Functor-Applicative-Monad_Proposal
for more information.